requirements.txt
flask
flask_sqlalchemy
flask_mysqldb 
Flask-Mail2
flask_login
flask_socketio
mysql-connector
requests
pandas 
joblib
nltk
beautifulsoup4
docx
python-docx
PyPDF2
pyttsx3
scikit-learn
flask-session
python-dotenv